% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_}
\alias{geom_}
\alias{geom_.character}
\alias{geom_.ggbuilder_layer}
\alias{geom_.ggbuilder_layer_spec}
\alias{geom_.ggbuilder_plot_data}
\title{Pipable ggplot2 geoms}
\usage{
geom_(x, ...)

\method{geom_}{character}(x, ...)

\method{geom_}{ggbuilder_layer}(x, ...)

\method{geom_}{ggbuilder_layer_spec}(x, ...)

\method{geom_}{ggbuilder_plot_data}(x, ...)
}
\arguments{
\item{x}{One of:
\itemize{
\item A string: The suffix of a \code{geom_XXX} function used to create this geom.
\item A \pkg{ggbuilder} \link{layer}, \link{layer_spec}, or \link{plot_data} object representing
earlier data and data transformations used as input to this geom. In this
case, the second argument should be the suffix of a \code{geom_XXX} function
used to create this geom.
}}

\item{...}{arguments passed on to the geom's constructor.}
}
\value{
A ggbuilder \link{layer}, which is also a \link[ggplot2:layer]{ggplot2::layer}, and which can be
added to a \code{\link[=ggplot]{ggplot()}} object or piped into other ggbuilder functions.
}
\description{
A \code{\link[ggplot2:layer]{ggplot2::layer()}} containing a \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}; intended for piping.
}
\details{
\code{geom_()} can be used to replace typical \code{geom_XXX()} calls in \pkg{ggplot2}
code to make them pipable by using the \code{XXX} suffix as a string as the first
argument to \code{geom_()}.

For example, \code{geom_boxplot(...)} would become \code{geom_("boxplot", ...)}.
The primary difference between the former and the latter is that the latter
is both a \link[ggplot2:layer]{ggplot2::layer} \emph{and} a ggbuilder \link{layer} containing a \link{layer_spec}.
The contained \link{layer_spec} stores information on how to construct the
geom, which allows it to be combined with other objects (e.g. by piping it
into \code{remap()}) to construct more complex layers. See \emph{Examples}.
}
\examples{
library(ggplot2)

set.seed(123456)
df = data.frame(condition = c("A", "B", "C"), response = round(rnorm(30, 1:3), 1))

df |>
  ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  stat_("boxplot", aes(y = response)) |>
    geom_("label", aes(y = ymax, label = ymax)) |>
    remap(aes(fill = scales::alpha(color, 0.1))) +
  theme_light()

}
