% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbuilder-package.R
\docType{package}
\name{ggbuilder-package}
\alias{ggbuilder-package}
\alias{ggbuilder}
\title{A Data Flow Pipeline Approach to Building ggplot2 Layers}
\description{
\pkg{ggbuilder} provides a pipeline-based approach for building individual
\pkg{ggplot2} layers, analogous to \pkg{dplyr}-like data transformation pipelines.
}
\details{
As an alternative to \pkg{ggplot2} functions for controlling aesthetic evaluation
like \code{\link[=stage]{stage()}}, \pkg{ggbuilder} allows you to use function piping (i.e. \code{\link{|>}})
to chain data transformations, \link[ggplot2:ggplot2-ggproto]{ggplot2::Stat}s, and \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}s together
in order to construct layers.
\itemize{
\item See \code{\link[=geom_]{geom_()}} and \code{\link[=stat_]{stat_()}} for how to construct and chain geoms and stats
using \pkg{ggbuilder}
\item See \code{\link[=plot_data]{plot_data()}} for how to add data transformations into a layer definition
\item See \code{\link[=remap]{remap()}} for how to apply transformations to aesthetics after scales
have been calculated (analogous to \code{\link[ggplot2:aes_eval]{ggplot2::after_scale()}}).
}
}
