% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\alias{new_layer}
\alias{new_layer.ggbuilder_layer}
\alias{new_layer.ggbuilder_layer_spec}
\title{Pipable ggplot2 layers}
\usage{
new_layer(x)

\method{new_layer}{ggbuilder_layer}(x)

\method{new_layer}{ggbuilder_layer_spec}(x)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \pkg{ggbuilder} \link{layer_spec}
\item A \pkg{ggbuilder} \link{layer}
}}
}
\value{
A \pkg{ggbuilder} \link{layer}, which is also a \link[ggplot2:layer]{ggplot2::layer}, and which can be
added to a \code{\link[=ggplot]{ggplot()}} object or piped into other \pkg{ggbuilder} functions.
}
\description{
A \link[ggplot2:layer]{ggplot2::layer} built from a \link{layer_spec} and intended for
piping. Typically constructed using \code{\link[=geom_]{geom_()}} or \code{\link[=stat_]{stat_()}}, not this function.
}
\examples{

# See geom_() or stat_() for examples; this function
# should not usually be used directly.

}
