% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_data}
\alias{plot_data}
\title{Pipable ggplot2 layer data}
\usage{
plot_data(data = NULL)
}
\arguments{
\item{data}{One of:
\itemize{
\item \code{NULL}: Use the data from the \code{\link[=ggplot2]{ggplot2()}} object.
\item A function: apply this function to the data from the  \code{\link[=ggplot2]{ggplot2()}} object
when constructing the data for the layer.
\item A \code{\link[=data.frame]{data.frame()}}: use this data frame as the data for the layer.
}}
}
\value{
A ggbuilder \link{plot_data}, which can be piped into other ggbuilder
functions. Can also be piped into \pkg{dplyr} and \pkg{tidyr} verbs in order
to construct a new \link{plot_data} that will apply those verbs to the data.
}
\description{
A wrapper around data to be passed to the \code{data} argument of \code{\link[ggplot2:layer]{ggplot2::layer()}}.
Pipe it into a \pkg{ggbuilder} \link{layer} to set the data used by that layer.
}
\examples{

library(ggplot2)

set.seed(123456)
df = data.frame(condition = c("A", "B", "C"), response = round(rnorm(30, 1:3), 1))

df |>
ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  plot_data(\(x) x[x$condition \%in\% c("B", "C"),]) |>
    stat_("boxplot", aes(y = response)) |>
    geom_("label", aes(y = ymax, label = ymax))

df |>
  ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  plot_data() |>
    dplyr::filter(condition \%in\% c("B", "C")) |>
    dplyr::group_by(condition) |>
    dplyr::slice_max(response) |>
    geom_("label", aes(label = response))

}
