% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap.R
\name{remap}
\alias{remap}
\alias{remap.ggbuilder_layer}
\alias{remap.ggbuilder_layer_spec}
\title{Remap aesthetics after scales are applied}
\usage{
remap(x, mapping)

\method{remap}{ggbuilder_layer}(x, mapping)

\method{remap}{ggbuilder_layer_spec}(x, mapping)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \pkg{ggbuilder} \link{layer}
\item A \pkg{ggbuilder} \link{layer_spec}
}}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or \code{\link[=aes_]{aes_()}}.}
}
\value{
A ggbuilder \link{layer} with aesthetics provided in \code{mapping} applied after
scales are computed, using \code{\link[ggplot2:aes_eval]{ggplot2::after_scale()}}
}
\description{
Add this to the end of a \pkg{ggbuilder} pipeline to remap layer aesthetics
after the scale functions have been applied. Analogous to \code{\link[ggplot2:aes_eval]{ggplot2::after_scale()}}.
}
\examples{

library(ggplot2)

set.seed(123456)
df = data.frame(condition = c("A", "B", "C"), response = round(rnorm(30, 1:3), 1))

df |>
  ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  stat_("boxplot", aes(y = response)) |>
    geom_("label", aes(y = ymax, label = ymax)) |>
    remap(aes(fill = scales::alpha(color, 0.1))) +
  theme_light()

}
