% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R, R/plot-data.R
\name{geom_}
\alias{geom_}
\alias{geom_.character}
\alias{geom_.ggbuilder_layer}
\alias{geom_.ggbuilder_layer_spec}
\alias{geom_.ggbuilder_plot_data}
\alias{plot_data}
\title{Pipable ggplot2 geoms}
\usage{
geom_(x, ...)

\method{geom_}{character}(x, ...)

\method{geom_}{ggbuilder_layer}(x, ...)

\method{geom_}{ggbuilder_layer_spec}(x, ...)

\method{geom_}{ggbuilder_plot_data}(x, ...)

plot_data(data = NULL)
}
\arguments{
\item{x}{One of:
\itemize{
\item A string: The suffix of a \code{geom_XXX} function used to create this geom.
\item A \pkg{ggbuilder} \link{layer}, \link{layer_spec}, or \link{plot_data} object representing
earlier data and data transformations used as input to this geom. In this
case, the second argument should be the suffix of a \code{geom_XXX} function
used to create this geom.
}}

\item{...}{arguments passed on to the geom's constructor.}

\item{data}{One of:
\itemize{
\item \code{NULL}: Use the data from the \code{\link[=ggplot2]{ggplot2()}} object.
\item A function: apply this function to the data from the  \code{\link[=ggplot2]{ggplot2()}} object
when constructing the data for the layer.
\item A \code{\link[=data.frame]{data.frame()}}: use this data frame as the data for the layer.
}}
}
\value{
A \pkg{ggbuilder} \link{layer}, which is also a \link[ggplot2:layer]{ggplot2::layer}, and which can be
added to a \code{\link[=ggplot]{ggplot()}} object or piped into other \pkg{ggbuilder} functions.

A \pkg{ggbuilder} \link{plot_data}, which can be piped into other \pkg{ggbuilder}
functions. Can also be piped into \pkg{dplyr} and \pkg{tidyr} verbs in order
to construct a new \link{plot_data} that will apply those verbs to the data.
}
\description{
A \code{\link[ggplot2:layer]{ggplot2::layer()}} containing a \link[ggplot2:ggplot2-ggproto]{ggplot2::Geom}; intended for piping.

A wrapper around data to be passed to the \code{data} argument of \code{\link[ggplot2:layer]{ggplot2::layer()}}.
Pipe it into a \pkg{ggbuilder} \link{layer} to set the data used by that layer.
}
\examples{

library(ggplot2)

set.seed(123456)
df = data.frame(condition = c("A", "B", "C"), response = round(rnorm(30, 1:3), 1))

df |>
  ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  stat_("boxplot", aes(y = response)) |>
    geom_("label", aes(y = ymax, label = ymax)) |>
    remap(aes(fill = scales::alpha(color, 0.1))) +
  theme_light()


library(ggplot2)

set.seed(123456)
df = data.frame(condition = c("A", "B", "C"), response = round(rnorm(30, 1:3), 1))

df |>
ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  plot_data(\(x) x[x$condition \%in\% c("B", "C"),]) |>
    stat_("boxplot", aes(y = response)) |>
    geom_("label", aes(y = ymax, label = ymax))

df |>
  ggplot(aes(x = condition, y = response, color = condition)) +
  geom_boxplot() +
  plot_data() |>
    dplyr::filter(condition \%in\% c("B", "C")) |>
    dplyr::group_by(condition) |>
    dplyr::slice_max(response) |>
    geom_("label", aes(label = response))

}
